Complete stock_picker algorithm
Utilized nested loops that track minimum price and maximum difference.
Inner loops is only accessed if the price being evaluated is less than
the previously calculated minimum. Values are then only overwritten in
the inner loop if the difference between prices is greater than
previously calculated max. difference and if the index of the buy price
is lower than the index of the sell price. Function returns an array
containing the index indicating the best day to buy followed by best
day to sell.