Implement recursive bubble sort with swap tracker
Created a recursive bubble sort method that takes array as an input,
compares all adjacent values and swaps the first value with the second
if necessary. If any values are swapped, swap tracker is flagged.

After one pass through the array, the method is called recursively on
the entire array if the swap tracker was flagged. If not, the numbers
are already in order and the ordered array is returned.

Swap tracking adds a minimal degree of efficiency. Method could be
marginally more efficient if only n-1 values were compared in each
recursive call of bubble_sort (because we know the highest value made
it to n). Could not figure out how to keep track of the full, sorted
array without adding an additional parameter to bubble_sort. Decided
not to do this since the assignment only shows a method that accepts
one parameter.